Add integration test to btc_kyt_canister
